security-and-compliance-account/requirements.txt
aws-cdk.core==1.59.0
aws_cdk.aws-codecommit==1.59.0
aws_cdk.aws-codebuild==1.59.0
aws_cdk.aws-codepipeline==1.59.0
aws_cdk.aws-codepipeline-actions==1.59.0
aws-cdk.aws-s3==1.59.0
aws-cdk.aws-s3-notifications==1.59.0
aws-cdk.aws-cloudformation==1.59.0
aws-cdk.aws-lambda==1.59.0
aws-cdk.aws-lambda-event-sources==1.59.0
aws-cdk.aws-iam==1.59.0
aws-cdk.aws-kms==1.59.0
pylint